This feature is now part of Texpattern >= 4.7.0
As such, mark this project as discontinued.

This last commit updates the project template
just to sun set it off with a clean shave.

Also fixes the PR bugs while at it
for good measure.

While this never had a stable release, the
repository is kept around so that potential
existing installs of the pacakge do not
break, and for history reasons. May some
find any use for this or its code base.

Closes #1 #2